Added ProductModel
